rename the storage settings handler
